+++ tests/Util/fixtures/reader-closure-invalid-04.php
@@ -3,10 +3,9 @@
 declare(strict_types=1);
 
 use FormatPHP\ConfigInterface;
-use FormatPHP\Intl\LocaleInterface;
 use FormatPHP\MessageCollection;
 
-return function (ConfigInterface $config, $data, LocaleInterface $localeResolved): MessageCollection
+return function (ConfigInterface $config, $data): MessageCollection
 {
     return new MessageCollection();
 };