+++ tests/Util/fixtures/reader-closure-invalid-05.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 use FormatPHP\ConfigInterface;
 use FormatPHP\MessageCollection;
 
-return function (ConfigInterface $config, array $data, $localeResolved): MessageCollection
+return function (ConfigInterface $config, array $data)
 {
     return new MessageCollection();
 };